// shipping/src/modules/shipping-app/src/interfaces/products/collection.interface.ts
import { FormattedFilter } from "./filter.interface";
import { PaginationInfoCamelCase } from "./pagination.interface";

export const mimeTypes: string = 'png|jpg|jpeg|bmp';

export interface Collection {
  _id?: string;
  name: string;
  slug: string;
  description?: string;
  image?: string;
  channelSets: string[];
  activeSince: Date;
  activeTill?: Date;
  automaticFillConditions?: {
    strict: boolean,
    filters: FormattedFilter[]
  };
}

export interface CollectionsLoadedInterface {
  info: PaginationInfoCamelCase;
  collections: Collection[];
}